+++ examples/swap/program/src/curve.rs
@@ -29,50 +29,3 @@ fn calculate_rate_adjustment(token_a_amount: u64, token_b_amount: u64) -> u64 {
     }
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_calculate_swap_amount() {
-        // Test case 1: Balanced pool with small swap impact
-        let token_a_reserve = 2000;
-        let token_b_reserve = 2000;
-        let input_amount = 100;
-
-        let output = calculate_swap_amount(token_a_reserve, token_b_reserve, input_amount);
-        assert_eq!(output, 96);
-
-        // Test case 2: Unbalanced pool favoring token_a
-        let token_a_reserve = 3000;
-        let token_b_reserve = 1000;
-        let input_amount = 100;
-
-        let output = calculate_swap_amount(token_a_reserve, token_b_reserve, input_amount);
-        assert_eq!(output, 91);
-
-        // Test case 3: Large swap impact on a balanced pool
-        let token_a_reserve = 1000;
-        let token_b_reserve = 1000;
-        let input_amount = 500;
-
-        let output = calculate_swap_amount(token_a_reserve, token_b_reserve, input_amount);
-        assert_eq!(output, 334);
-
-        // Test case 4: Minimal swap impact on a large pool
-        let token_a_reserve = 10000;
-        let token_b_reserve = 10000;
-        let input_amount = 10;
-
-        let output = calculate_swap_amount(token_a_reserve, token_b_reserve, input_amount);
-        assert_eq!(output, 10);
-
-        // Test case 5: Unbalanced pool favoring token_b
-        let token_a_reserve = 1000;
-        let token_b_reserve = 3000;
-        let input_amount = 100;
-
-        let output = calculate_swap_amount(token_a_reserve, token_b_reserve, input_amount);
-        assert_eq!(output, 114);
-    }
-}
